Fix for a CLN related bug on 32 bit systems. Integer((1<<29)+1) and Integer((long int)((1<<29)+1)) gave different values.  This was confirmed on vm-int1.cims.nyu.edu. See http://www.ginac.de/CLN/cln_3.html#SEC15 for more details. rational_white and integer_white have tests covering this.

+++ src/util/rational_cln_imp.h
@@ -125,8 +125,8 @@ public:
   /**
    * Constructs a canonical Rational from a numerator.
    */
-  Rational(signed int n) : d_value(n) { }
-  Rational(unsigned int n) : d_value(n) { }
+  Rational(signed int n) : d_value((signed long int)n) { }
+  Rational(unsigned int n) : d_value((unsigned long int)n) { }
   Rational(signed long int n) : d_value(n) { }
   Rational(unsigned long int n) : d_value(n) { }
 
@@ -138,10 +138,10 @@ public:
   /**
    * Constructs a canonical Rational from a numerator and denominator.
    */
-  Rational(signed int n, signed int d) : d_value(n) {
+  Rational(signed int n, signed int d) : d_value((signed long int)n) {
     d_value /= d;
   }
-  Rational(unsigned int n, unsigned int d) : d_value(n) {
+  Rational(unsigned int n, unsigned int d) : d_value((unsigned long int)n) {
     d_value /= d;
   }
   Rational(signed long int n, signed long int d) : d_value(n) {
